+++ packages/core/src/tools/index.ts
@@ -1,4 +1,5 @@
 export { generateSites, GenerateSitesConfig } from './generateSites';
 export { generateMetadata } from './generateMetadata';
 export { scaffoldComponent } from './scaffold';
+export { GenerateMapFunction, GenerateMapArgs } from './generate-map';
 export * from './templating';