+++ packages/react/src/components/Form.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef, useState } from 'react';
 import { ComponentRendering } from '@sitecore-content-sdk/core/layout';
 import { form } from '@sitecore-content-sdk/core';
-import { useSitecoreContext } from '../enhancers/withSitecoreContext';
+import { useSitecore } from '../enhancers/withSitecore';
 
 let { executeScriptElements, loadForm, subscribeToFormSubmitEvent } = form;
 
@@ -41,10 +41,10 @@ export const Form = ({ params, rendering }: FormProps) => {
   const id = params?.RenderingIdentifier;
   const [error, setError] = useState(false);
   const [content, setContent] = useState('');
-  const context = useSitecoreContext();
+  const context = useSitecore();
   const formRef = useRef<HTMLDivElement>(null);
 
-  const isEditing = context.sitecoreContext.pageEditing;
+  const isEditing = context.pageContext.pageEditing;
 
   useEffect(() => {
     if (!content) {
@@ -53,7 +53,7 @@ export const Form = ({ params, rendering }: FormProps) => {
         .catch(() => {
           if (isEditing) {
             console.error(
-              `Failed to load form with id ${params.FormId}. Check debug logs for sitecore-jss:form for more details.`
+              `Failed to load form with id ${params.FormId}. Check debug logs for content-sdk:form for more details.`
             );
           }
           setError(true);
@@ -71,7 +71,9 @@ export const Form = ({ params, rendering }: FormProps) => {
   if (isEditing) {
     if (error) {
       return (
-        <div className="sc-jss-placeholder-error">There was a problem loading this section</div>
+        <div className="sc-content-sdk-placeholder-error">
+          There was a problem loading this section
+        </div>
       );
     }
   }